+++ primefaces/src/main/frontend/packages/core/src/core/core.ts
@@ -1128,7 +1128,7 @@ export class Core {
         }).on('pfAjaxComplete' + namespace, function(e, xhr, settings, args) {
             if (isXhrSource.call(this, widget, settings)) {
                 widget.ajaxCount--;
-                if (widget.ajaxCount > 0 || !args || args.redirect) {
+                if (widget.ajaxCount > 0 || args?.redirect) {
                     return;
                 }