+++ primefaces/src/main/java/org/primefaces/component/selectbooleanbutton/SelectBooleanButton.java
@@ -66,6 +66,13 @@ public class SelectBooleanButton extends SelectBooleanButtonBaseImpl {
             styleClass = styleClass + " ui-state-error";
         }
 
+        if ("small".equals(getSize())) {
+            styleClass = styleClass + " ui-button-sm";
+        }
+        else if ("large".equals(getSize())) {
+            styleClass = styleClass + " ui-button-lg";
+        }
+
         String userStyleClass = getStyleClass();
         if (userStyleClass != null) {
             styleClass = styleClass + " " + userStyleClass;