+++ primefaces/src/main/java/org/primefaces/component/selectbooleanbutton/SelectBooleanButtonBase.java
@@ -82,4 +82,7 @@ public abstract class SelectBooleanButtonBase extends HtmlSelectBooleanCheckbox 
     @Property(description = "The aria-label attribute is used to define a string that labels the current element for accessibility.")
     public abstract String getAriaLabel();
 
+    @Property(description = "Defines the size of the button, valid values are \"small\" and \"large\".")
+    public abstract String getSize();
+
 }