+++ docs/16_0_0/gettingstarted/whatsnew.md
@@ -8,6 +8,8 @@ This page contains a list of big features. Please check the GitHub issues for al
 
 Look into [migration guide](https://primefaces.github.io/primefaces/16_0_0/#/../migrationguide/16_0_0) for more enhancements and changes.
 
+* Button/CommandButton/LinkButton/MenuButton/SelectBooleanButton/SelectManyButton/SelectOneButton/SplitButton
+    * Added `size` property to control the dimensions of the component, possible values are `small` and `large`
 * ColumnToggler
     * Added `showSelectAll` property to show/hide the select all checkbox
 * Dialog/ConfirmDialog